Forgot to move a few more things into EmptyCase
